Refactor view methods
